Refactor codes & fix stylelint errors
