+++ src/index.html
@@ -24,7 +24,7 @@
 
             <nav class="header-content-nav" id="menu">
                 <ul class="nav-content">
-                    <li><a href="#">All Movies</a></li>
+                    <li><a href="#" class="total-movies">All Movies()</a></li>
                     <li><a href="#">Feedbacks</a></li>
                     <li><a href="#">Contact-Us</a></li>
