+++ src/index.js
@@ -1,7 +1,6 @@
 import './style.css';
 import 'font-awesome/css/font-awesome.min.css';
-import initializeApp from './modules/none';
-
+import initializeApp from './modules/display.js';
 
 window.addEventListener('DOMContentLoaded', initializeApp);
 // import image from './asset/img/14.png';
