+++ src/modules/display.js
@@ -1,224 +1,93 @@
-// import { fetchShows, fetchShowsDetails } from './fetchShows.js';
-// import { fetchComment, postComment } from './fetchComment.js';
-// import { getLikes, addLikes } from './fetchLikes.js';
-// import { filter, update } from 'lodash';
+import { fetchShows } from './fetchShows.js';
+import { getLikes, addLikes } from './fetchLikes.js';
+import addEventToCommentButton from './popUp.js';
 
-// const popUpModal = document.querySelector('.popUp-modal');
-// const movieContainer = document.querySelector('.container-cards');
-// const url = 'https://api.tvmaze.com/shows';
-// const commentUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/TmorUv6CAxzfjMnV7ubN';
-// const likesUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/TmorUv6CAxzfjMnV7ubN/likes';
-// const displayPopUp = async (buttonId) => {
-//   try {
-//     const showsPopDetails = await fetchShowsDetails(url, buttonId);
-//     // console.log(showsPopDetails);
-//     const comments = await fetchComment(commentUrl, buttonId);
-//     const commentLength = comments.length;
-//     const elements = comments.map((comment) => {
-//       const li = `<li class="p-class">
-//                 <span>${comment.creation_date}</span>
-//                 <span>${comment.username}</span>
-//                 <span>${comment.comment}</span>
-//             </li>`;
-//       return li;
-//     });
-//     const showPopObject = {
-//       id: showsPopDetails.id,
-//       title: showsPopDetails.name,
-//       image: showsPopDetails.image.medium,
-//       description: showsPopDetails.summary,
-//       genres: showsPopDetails.genres,
-//       language: showsPopDetails.language,
-//       rating: showsPopDetails.rating.average,
-//     };
-//     // console.log(showPopObject);
-//     const popUpElement = `
+const url = 'https://api.tvmaze.com/shows';
+const likesUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/TmorUv6CAxzfjMnV7ubN/likes';
 
-//     <button class="close-button"><i class="fa fa-close"></i></button>
+// Function to process show data by merging it with likes
+const processShowData = (movies, likes) => movies.map((show) => {
+  const itemlike = likes.find((like) => like.item_id === show.id);
+  const likesCount = itemlike?.likes || 0;
 
-//     <h3 class="m-title">${showPopObject.title}</h3>
+  return {
+    id: show.id,
+    image: show.image.medium,
+    title: show.name,
+    likes: likesCount,
+  };
+});
 
-//         <div class="imgdes">
-//         <img src="${showPopObject.image}" alt="" class="popUp-img">
-//         <p  class="para">${showPopObject.description}</p>
-//         </div>
+// Function to generate HTML elements for shows
+const movieElements = (show) => {
+  const movieContentContainer = document.createElement('div');
+  movieContentContainer.classList.add('movie-content-container');
+  movieContentContainer.innerHTML = `      
+        <div class="movie-thumbnail">
+          <img src="${show.image}" alt="" class="movie-shows" id="${show.id}">
+          <div class="overlay"> <i class="fa fa-play"></i></div>
+        </div>
+        <div class="movie-content">
+          <h3 class="movie-title">${show.title} </h3>
+          <div class="movie-comment">
+            <div class="show-likes">
+              <i class="fa fa-heart" data-id="${show.id}"></i>
+              <p class="show-likes-count">
+                <span class="likes-count liked" id="${show.id}"><span>${show.likes}</span> like </span>
+              </p>
+            </div>
+            <button class="comment-button" data-index="${show.id}">Comment</button>
+            <button class="comment-button1">  <i class="fa fa-comment" data-index1="${show.id}"></i></button>
+          </div>
+        </div>
+    `;
+  return movieContentContainer;
+};
 
-//         <div class="modal-content">
-           
-//             <div class="descriptions">
-//             <p>${showPopObject.genres}</p>
-//             <p>${showPopObject.language}</p>
-//             <p>${showPopObject.rating}</p>
-//              </div>
-//        </div>
-//                 <div class="commentcontainer">
+const renderMovies = (movies) => {
+  const movieContainer = document.querySelector('.container-cards');
+  movieContainer.innerHTML = '';
+  movies.forEach((movie) => {
+    const movieElement = movieElements(movie);
+    movieContainer.appendChild(movieElement);
+  });
+};
 
-//               <h2 class="popUp-h2-comment-title">Comments(2)</h2>
-//             <div class="comment-container">
-          
-//             <ul class="popUP-comment-content">
-//                 ${elements.join('')}
-//             </ul>
-//         </div>
-//         </div>
-              
-      
-//         <form action="" class="comment-form">
-//             <input type="text" name="text" id="input-name" required placeholder="Your name">
-//             <textarea name="text" id="comment-text" cols="30" rows="10" required placeholder="Add a comment" ></textarea>
-//             <button type="submit" id="${showPopObject.id}" class="submit-button">Submit</button>
-//         </form>
+// Function to update likes
+const updateLikesCount = (id) => {
+  const likesCount = document.querySelector(`.likes-count[id="${id}"] > span`);
+  const likesCountValue = parseInt(likesCount.textContent, 10);
+  likesCount.textContent = `${likesCountValue + 1}`;
+};
 
-//     `;
+// Function to handle like button click
+const handleLikeButtonClick = async (heartbutton) => {
+  const heartbuttonId = parseInt(heartbutton.dataset.id, 10);
+  const response = await addLikes(likesUrl, heartbuttonId);
+  if (response.status === 201) {
+    updateLikesCount(heartbuttonId);
+  }
+};
 
-//     popUpModal.innerHTML = popUpElement;
-//     const popUpCountComment = document.querySelectorAll('.popUp-h2-comment-title');
-//     popUpCountComment.forEach((element) => {
-//       element.textContent = `Comments ${commentLength}`;
-//     });
-//     const closeButton = document.querySelector('.close-button');
-//     closeButton.addEventListener('click', () => {
-//       popUpModal.style.display = 'none';
-//     });
-//     const submitButton = document.querySelector('.submit-button');
-//     const commentForm = document.querySelector('.comment-form');
-//     submitButton.addEventListener('click', async (e) => {
-//       e.preventDefault();
-//       const submitButtonId = e.target.id;
-//       const inputName = commentForm.querySelector('#input-name');
-//       const commentText = commentForm.querySelector('#comment-text');
-//       //  eslint-disable-next-line camelcase
-//       const item_id = submitButtonId;
-//       const username = inputName.value;
-//       const comment = commentText.value;
-//       const userInfo = { item_id, username, comment };
-//       try {
-//         const response = await postComment(commentUrl, userInfo);
-//         if (response === 'Created') {
-//           displayPopUp(submitButtonId);
-//           inputName.value = '';
-//           commentText.value = '';
-//         }
-//       } catch (error) {
-//         return error;
-//       }
-//       return userInfo;
-//     });
-//     return showPopObject;
-//   } catch (error) {
-//     return error;
-//   }
-// };
-// // eslint-disable-next-line import/prefer-default-export
-// export const displayShows = async () => {
-//   movieContainer.innerHTML = '';
-//   try {
-//     const shows = await fetchShows(url);
-//     // eslint-disable-next-line max-len
-//     const filteredShowObject = shows.map((show) => {
-//       const showObj = {
-//         id: show.id,
-//         image: show.image.medium,
-//         title: show.name,
-//       };
-//       return showObj;
-//     });
-//     const createShowElement = filteredShowObject.map((filterShow) => {
-//       const showElement = `<div class="movie-content-container">
-//                     <div class="movie-thumbnail">
-//                         <img src="${filterShow.image}" alt="" class="movie-shows" id="${filterShow.id}">
-//                         <div class="overlay"> <i class="fa fa-play"></i></div>
-//                     </div>
-//                     <div class="movie-content">
-//                         <h3 class="movie-title">${filterShow.title} </h3>
-//                         <div class="movie-comment">
-//                         <div class="show-likes">
-//                         <i class="fa fa-heart" data-id="${filterShow.id}"></i>
-//                             <p class="show-likes-count">
-//                               <span class="likes-count liked" id=${filterShow.id}><span>0</span> like </span>
-                          
-//                             </p>
-//               </div>
-                       
-//                         <button class="comment-button" data-index="${filterShow.id}">Comment</button>
-//                         <button class="comment-button1">  <i class="fa fa-comment" data-index1="${filterShow.id}"></i></button>
-//                     </div>
-//                     </div>
+const addEventToLikeButtons = () => {
+  const likeButtons = document.querySelectorAll('.fa-heart');
+  likeButtons.forEach((heartbutton) => {
+    heartbutton.addEventListener('click', (e) => {
+      handleLikeButtonClick(e.currentTarget);
+    });
+  });
+};
 
-//                 </div>`;
-//       return showElement;
-//     }).join('');
+// Initialize the application
+const initializeApp = async () => {
+  const movies = await fetchShows(url);
+  const likes = await getLikes(likesUrl);
+  const filteredShowObject = processShowData(movies, likes);
+  const totalMovies = document.querySelector('.total-movies');
+  totalMovies.textContent = ` All movies (${filteredShowObject.length})`;
+  renderMovies(filteredShowObject);
+  addEventToLikeButtons();
+  addEventToCommentButton();
+};
 
-//     const popupPosition = (button, popUpModal, e) => {
-//       const commentClick = button.getBoundingClientRect();
-
-//       const left = e.clientX - commentClick.left;
-//       const top = e.clientY - commentClick.top;
-
-//       popUpModal.style.left = `${left}px`;
-//       popUpModal.style.top = `${top}px`;
-//     };
-
-//     movieContainer.insertAdjacentHTML('beforeend', createShowElement);
-//     const commentButton = document.querySelectorAll('.comment-button');
-//     commentButton.forEach((button) => {
-//       button.addEventListener('click', (e) => {
-//         const buttonId = parseInt(e.target.dataset.index, 10);
-//         displayPopUp(buttonId);
-//         popupPosition(button, popUpModal, e);
-
-//         popUpModal.style.display = 'flex';
-//       });
-//     });
-//     const commentButton1 = document.querySelectorAll('.comment-button1');
-//     commentButton1.forEach((button) => {
-//       button.addEventListener('click', (e) => {
-//         const buttonId = parseInt(e.target.dataset.index1, 10);
-//         displayPopUp(buttonId);
-
-//         popupPosition(button, popUpModal, e);
-
-//         popUpModal.style.display = 'flex';
-//       });
-//     });
-
-//     const toggleHeartColor = (event) => {
-//       const heartIcon = event.currentTarget;
-//       const currentColor = window.getComputedStyle(heartIcon).color;
-
-//       if (currentColor !== 'rgb(255, 255, 255)') {
-//         heartIcon.style.color = 'white';
-//       } else {
-//         heartIcon.style.color = 'red';
-//       }
-//     };
-
-//     document.querySelectorAll('.fa-heart').forEach((heartIcon) => {
-//       heartIcon.addEventListener('click', toggleHeartColor);
-//     });
-
-//     document.querySelectorAll('.fa-heart').forEach((heartbutton) => {
-//       heartbutton.addEventListener('click', async (e) => {
-//         const parent = e.currentTarget.parentElement;
-//         console.log(parent);
-//         const countSpan = parent.querySelector('.likes-count');
-//         console.log(countSpan);
-//         let likesCount = countSpan.textContent ? parseInt(countSpan.textContent, 10) : 0;
-//         console.log(likesCount);
-//         if (parent.classList.contains('liked')) {
-//           parent.classList.remove('liked');
-//           likesCount -= 1;
-//           // const addLikes = addLikes(likesUrl, target);
-//           countSpan.textContent = `${likesCount} like`;
-//         } else {
-//           parent.classList.add('liked');
-//           likesCount += 1;
-//           countSpan.textContent = `${likesCount} like`;
-//         }
-//       });
-//     });
-//     return createShowElement;
-//   } catch (error) {
-//     return error;
-//   }
-// };
+export default initializeApp;
