+++ src/modules/popUp.js
@@ -1,4 +1,4 @@
-import {  fetchShowsDetails } from './fetchShows.js';
+import { fetchShowsDetails } from './fetchShows.js';
 import { fetchComment, postComment } from './fetchComment.js';
 
 const popUpModal = document.querySelector('.popUp-modal');
@@ -6,121 +6,14 @@ const popUpModal = document.querySelector('.popUp-modal');
 const url = 'https://api.tvmaze.com/shows';
 const commentUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/TmorUv6CAxzfjMnV7ubN';
 
-// const displayPopUp1 = async (buttonId) => {
-//   try {
-//     const showsPopDetails = await fetchShowsDetails(url, buttonId);
-//     // console.log(showsPopDetails);
-//     const comments = await fetchComment(commentUrl, buttonId);
-//     const commentLength = comments.length;
-//     const elements = comments.map((comment) => {
-//       const li = `<li class="p-class">
-//                 <span>${comment.creation_date}</span>
-//                 <span>${comment.username}</span>
-//                 <span>${comment.comment}</span>
-//             </li>`;
-//       return li;
-//     });
-//     const showPopObject = {
-//       id: showsPopDetails.id,
-//       title: showsPopDetails.name,
-//       image: showsPopDetails.image.medium,
-//       description: showsPopDetails.summary,
-//       genres: showsPopDetails.genres,
-//       language: showsPopDetails.language,
-//       rating: showsPopDetails.rating.average,
-//     };
-//     // console.log(showPopObject);
-//     const popUpElement = `
-
-//     <button class="close-button"><i class="fa fa-close"></i></button>
-
-//     <h3 class="m-title">${showPopObject.title}</h3>
-
-//         <div class="imgdes">
-//         <img src="${showPopObject.image}" alt="" class="popUp-img">
-//         <p  class="para">${showPopObject.description}</p>
-//         </div>
-
-//         <div class="modal-content">
-           
-//             <div class="descriptions">
-//             <p>${showPopObject.genres}</p>
-//             <p>${showPopObject.language}</p>
-//             <p>${showPopObject.rating}</p>
-//              </div>
-//        </div>
-//                 <div class="commentcontainer">
-
-//               <h2 class="popUp-h2-comment-title">Comments(2)</h2>
-//             <div class="comment-container">
-          
-//             <ul class="popUP-comment-content">
-//                 ${elements.join('')}
-//             </ul>
-//         </div>
-//         </div>
-              
-      
-//         <form action="" class="comment-form">
-//             <input type="text" name="text" id="input-name" required placeholder="Your name">
-//             <textarea name="text" id="comment-text" cols="30" rows="10" required placeholder="Add a comment" ></textarea>
-//             <button type="submit" id="${showPopObject.id}" class="submit-button">Submit</button>
-//         </form>
-
-//     `;
-
-//     popUpModal.innerHTML = popUpElement;
-//     const popUpCountComment = document.querySelectorAll('.popUp-h2-comment-title');
-//     popUpCountComment.forEach((element) => {
-//       element.textContent = `Comments ${commentLength}`;
-//     });
-
-//     const closeButton = document.querySelector('.close-button');
-//     closeButton.addEventListener('click', () => {
-//       popUpModal.style.display = 'none';
-//     });
-    
-//     const submitButton = document.querySelector('.submit-button');
-//     const commentForm = document.querySelector('.comment-form');
-//     submitButton.addEventListener('click', async (e) => {
-//       e.preventDefault();
-//       const submitButtonId = e.target.id;
-//       const inputName = commentForm.querySelector('#input-name');
-//       const commentText = commentForm.querySelector('#comment-text');
-//       //  eslint-disable-next-line camelcase
-//       const item_id = submitButtonId;
-//       const username = inputName.value;
-//       const comment = commentText.value;
-//       const userInfo = { item_id, username, comment };
-//       try {
-//         const response = await postComment(commentUrl, userInfo);
-//         if (response === 'Created') {
-//           displayPopUp(submitButtonId);
-//           inputName.value = '';
-//           commentText.value = '';
-//         }
-//       } catch (error) {
-//         return error;
-//       }
-//       return userInfo;
-//     });
-//     return showPopObject;
-//   } catch (error) {
-//     return error;
-//   }
-// };
 // Function to create a comment list HTML from an array of comments
-const createCommentList = (comments) => {
-  return comments.map((comment) => {
-    return `
+const createCommentList = (comments) => comments.map((comment) => `
       <li class="p-class">
         <span>${comment.creation_date}</span>
         <span>${comment.username}</span>
         <span>${comment.comment}</span>
       </li>
-    `;
-  }).join('');
-};
+    `).join('');
 
 // Function to create the pop-up modal HTML
 const createPopUpModal = (showDetails, comments) => {
@@ -162,7 +55,7 @@ const displayPopUp = async (buttonId) => {
   try {
     const showsPopDetails = await fetchShowsDetails(url, buttonId);
     const comments = await fetchComment(commentUrl, buttonId);
-    
+
     const showPopObject = {
       id: showsPopDetails.id,
       title: showsPopDetails.name,
@@ -203,6 +96,7 @@ const displayPopUp = async (buttonId) => {
           inputName.value = '';
           commentText.value = '';
         }
+        return response;
       } catch (error) {
         return error;
       }
@@ -214,9 +108,6 @@ const displayPopUp = async (buttonId) => {
   }
 };
 
-
-
-
 const popupPosition = (button, popUpModal, e) => {
   const commentClick = button.getBoundingClientRect();
 
@@ -227,7 +118,7 @@ const popupPosition = (button, popUpModal, e) => {
   popUpModal.style.top = `${top}px`;
 };
 
-const addEventToCommentButton = (button) => {
+const addEventToCommentButton = () => {
   const commentButton = document.querySelectorAll('.comment-button');
   commentButton.forEach((button) => {
     button.addEventListener('click', (e) => {
@@ -249,20 +140,6 @@ const addEventToCommentButton = (button) => {
       popUpModal.style.display = 'flex';
     });
   });
-}
-// const toggleHeartColor = (event) => {
-//   const heartIcon = event.currentTarget;
-//   const currentColor = window.getComputedStyle(heartIcon).color;
-
-//   if (currentColor !== 'rgb(255, 255, 255)') {
-//     heartIcon.style.color = 'white';
-//   } else {
-//     heartIcon.style.color = 'red';
-//   }
-// };
-
-// document.querySelectorAll('.fa-heart').forEach((heartIcon) => {
-//   heartIcon.addEventListener('click', toggleHeartColor);
-// });
+};
 
 export default addEventToCommentButton;
